facturas-06/10/2022

+++ database/migrations/2022_07_28_031636_create_proveedores_table.php
@@ -23,7 +23,7 @@ class CreateProveedoresTable extends Migration
             $table->string('poblacion', 50);
             $table->string('estado', 50);
             $table->string('pais', 50);
-            $table->integer('telefono')->unsigned();
+            $table->bigInteger('telefono')->unsigned();
             $table->string('email_proveedor', 50);
             $table->timestamps();
         });
